Reverse the primes in place in problem60.py
Using frontier.pop() will search primes in ascending order, as before.
The difference is that searching for child nodes will now happen in
descending order. This is what we want. After those children are each
appended to the stack, the next node to be popped off will be the one
with the smallest members.